refactor(frontend): make object for internal reduce of fetchSolTransactionsForSignature (#5059)
# Motivation

We refactor the `reduce` function inside
`fetchSolTransactionsForSignature` to return an object, so that we can
prepare for further calculation: we will calculate the cumulative
balance in another PR, to pass it to the mapping and calculate the
`closeAccount` value (from PR
https://github.com/dfinity/oisy-wallet/pull/5056).

# Changes

Instead of returning an array of parsed transactions `SolTransactionUi`,
the reduce function returns an object that is `{ parsedTransactions:
SolTransactionUi[] }`.

Nothing changed in the implementation of the functions.

# Tests

Current tests are enough.