feat(frontend): calculate cumulative balances when parsing Solana transactions (#5062)
# Motivation

The redeemed SOL value of a Solana `closeAccount` transaction is not
explicitly indicated in the transaction data. So it must be
inferred/calculated from the latest SOL balance of the ATA that is being
closed.

After PR https://github.com/dfinity/oisy-wallet/pull/5056 , we can
provide the accumulated balances to the mapper.

To calculate them we keep a record in the `reduce` function: every
instruction inside a generic transaction will affect the cumulative SOL
balance of each address.

It is guaranteed to be progressive since the RPC returns the list of
instructions in the correct internal chronological order.

The redeemed SOL value is the latest total SOL balance of the closed
Associated Token Account.

NOTE: We include WSOL in the calculation, because it is used to affect
the SOL balance of the ATA.


# Changes

<!-- List the changes that have been developed -->

# Tests

Apart from expanding the current tests, we verified with a [practical
transaction](https://solscan.io/tx/inspector?txhash=4E88TD8dpeivGbAn83DQcfQYyBcPF869Q9Nw9xfRvbZMDB46EMwC8FEqidJ5PoCZRSZVrBJJz486Ncju7duD3kwn)
comparing our results with Solscan.

![Screenshot 2025-03-06 at 22 40
33](https://github.com/user-attachments/assets/15332bc9-9f50-46e4-af1a-8b7bd597c31f)
![Screenshot 2025-03-06 at 22 41
56](https://github.com/user-attachments/assets/284bdc86-13f4-4724-a18f-f0401f1a4968)