feat(frontend): user-amount utils (#5076)
# Motivation

In this PR, the user-amount utils are created. They are gonna replace
the existing `convert.utils`. `validateUserAmount` is implemented the
way that it can be used in swap, send and convert flow.

In the next PR, I will remove the old file and update SwapForm and
ConvertAmountSource to start using this method.

# Changes

New util is created.

# Tests

Provided.