feat(build): save ckERC20 metadata (#5043)
# Motivation

Ideally we would like to have the ckERC20 metadata already in the code
base, to avoid fetching them at runtime, for each user. So, we duplicate
what we do for the SNS tokens and fetch and save the metadata in the
scripts that build the ckERC20 JSON file.

The next step will be to use such data and remove the fetching in the
code.

# Changes

- Create some utils for building tokens, for example to get metadata and
to create an agent.
- Create more types/schemas to have env ck ERC20 tokens with metadata
and icons.
- Fetch metadata when building the orchestrators info.
- Split the metadata that we need from the icons.
- Change method `saveTokenLogo` to simply use the data already fetched.

# Tests

I first injected a few changes in `tokens.ckerc20.json` and then run the
script on this PR.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>