feat(frontend): New bottom main navigation for mobile (#4955)
# Motivation

It has been decided by the design team that the mobile version of the
wallet needs a dedicated navigation bar.

Here we add a bottom navbar and adjust some aspects of the app to
accommodate that change.

# Changes

- Added new padding 24 variable and applied it on the footer when signed
in and on screen smaller than `md`
- Added mobile navigation wrapper component
- Relocated main navigation items to seperate component in order to
reuse the main items for all navs
- Added new color and tailwind class `bg-primary-inverted-alt` which is
used for bottom nav bg
- Added styling for new navbar
- Added new mobile nav to layout


# Tests

![image](https://github.com/user-attachments/assets/97d6fdbc-cf99-44b7-bab0-136d1fc38c39)

![image](https://github.com/user-attachments/assets/cdbedbed-e73f-4712-9c5f-1a386b927cf1)

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>
Co-authored-by: Max <bitdivine@users.noreply.github.com>