fix(e2e): added wait for carousel fade out functionality (#5101)
# Motivation

With recent changes to the carousel we also implemented a fade out
effect when swapping pages. For the e2e tests that sometimes resulted in
a screenshot where the carousel is still showing when it shouldnt.

# Changes

With this PR we implement a functionality for the relevant tests, so
that we always wait for the carousel to completely fade out.

# Tests

CI runs on this PR

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>