feat(build): use single package for Solana Web3.js v2 (#5116)
# Motivation

The package Solana Web3 v2 is still going through several change. Since
we are using the sub-packages of it, it is still not easily manageable.
Furthermore, we use small parts of such packages.

My proposal is to work directly with the bigger package
`@solana/web3.js` and chunk it.

# Note

Most recently they even changed the name of the main package (see
https://github.com/anza-xyz/kit/releases/tag/v2.1.0), but I will do that
in another PR.

# Changes

- Uninstall all `@solana/*` packages.
- Install `@solana/web3.js` v2.
- Adapt the imports.
- Add the packages to the lazy load options.

# Tests

Current tests and CI are sufficient.