feat(frontend): pass list of address and tokens to Solana transactions mapper (#5148)
# Motivation

We want to reduce the number of calls done to the Solana RPC. One way of
doing it is to keep track of the addresses in a certain signature and
their respective SPL token/mint. In this way, we can avoid calling
`getAccountInfo` more than once for the same addresses in the same
signature.

# Changes

- Pass a record of Solana addresses vs their token address (their mint)
to the Solana transaction mapper.
- Use the map in the Solana token sub-mapper: we will first check if the
token address is already associated for the source or the destination of
the mapped transaction.

# Tests

Added tests.