feat(frontend): remove the warning for non-ATA destination address (#5175)
# Motivation

The PRODSEC review suggested that we warn/inform the user when the
destination input for Solana is not an ATA address, since we will derive
it in our code.

This information can happen either in the form or in the confirmation
review, or both, or in whichever form it is acceptable to have the user
aware.

However, we are going to revisit how to show this information further
on. For the current release we temporarily remove it.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>