fix(frontend): Fix safe area on iOS standalone mode and iOS firefox (#5194)
# Motivation

The safe area CSS env rule seemed to show no effect for iOS standalone
mode (when added to home screen) and iOS firefox. This rule is needed to
account for the mobile dynamic navigation bars that can or cannot appear
depending on scroll etc.

# Changes

- Added `viewport-fit=cover` to the viewport meta tag, as this is what
should actually be set for the safe area env to be available
- Moved the CSS rule to the root of the mobile nav, instead of each
element independently

# Tests

- Tested on iPhone 14 max pro newest iOS Safari, Firefox, Chrome and
Standalone mode.
- Tested on iPhone 15 newest iOS Safari, Firefox and Chrome

Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>
Co-authored-by: Max <bitdivine@users.noreply.github.com>