feat(frontend): hide the destination for pending ERC20 transactions (#5206)
# Motivation

When we have a pending ERC20 transaction, we typically receive an
"not-totally-correct" information in the `to` field: it shows the ERC20
token (smart contract) address.

This could lead to confusions on the user side. For example the user
could copy such an address thinking it was the original destination
address, and re-use the copied address in a new transaction.

So, we avoid showing the `to` address for ERC20 pending transactions.  


![Screenshot 2025-03-13 at 08 41
55](https://github.com/user-attachments/assets/94259a5d-1914-41a7-9014-d8be7c9cab18)


Unchanged for ETH native token.

![Screenshot 2025-03-13 at 08 42
31](https://github.com/user-attachments/assets/7e99ddc9-491b-4632-9c3d-636d53c1be24)