feat(frontend): map CK minter contract name for ERC20 transactions (#5213)
# Motivation

We are already mapping the known ERC20 token names to display instead of
their addresses in a transaction detail. A further step is to map the CK
Minter too: if we know their address, we can map it and display a better
name instead of the address.

# Changes

- Export util `toCkMinterAddress`
- Add mapping of relevant CK minter addresses in util
`mapAddressToName`.
- Provide CK Minter Info to the caller of `mapAddressToName` in
component `EthTransactionModal`.

# Tests

Added tests.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>