fix(frontend): Fixate tailwind version to 4.0.9 (#5215)
# Motivation

Since the latest minor versions of tailwind introduce breaking changes
to gradient and viewport classes (affecting token hero and footer), we
fixate the version of tailwind to the last working minor version.

The last working minor version that still works well with gradients and
viewports is 4.0.9

# Changes

Changed tailwind version to 4.0.9 and removed `^` to avoid going to the
latest minor version.

# Tests

Working footer and gradient:

![image](https://github.com/user-attachments/assets/752a0341-c6b1-447a-8129-39c8e7795287)