feat(frontend): correctly handle ERC20 token addresses in transactions list/details (#5208)
# Motivation

This branch was created to release a patch with version `1.3.1`,
directly forking from version `1.3.0`.

The patch focuses on two points:
1. Visually improve how we show the Ethereum addresses in case it is a
known address (e.g. ERC20 tokens, CK minter). This helps the user to
copy or refer to the correct address in the transaction details.
2. Make the navigation bar position in the correct way in respect to
modals.

The PRs that were merged into this branch were peer-reviewed already:

- #5204
- #5205
- #5206 
- #5209 
- #5210 
- #5213 
- #5217 
- #5219 
- #5222 
- #5224

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: daviddecentage <80264282+daviddecentage@users.noreply.github.com>