test(frontend): simplify EthConvertTokenWizard tests (#5290)
# Motivation

The tests for `EthConvertTokenWizard` are flaky, because sometimes they
raise a console log for missing data (that is correct, since the tests
is done exactly for that).

However, we don't want console logs, we mock them before each test. So,
we do not reset the mocks but only clear them. And refactor the code to
make this work.

# Changes

- Move custom mocks in `beforeEach`.
- Substitute `resetAllMocks` with `clearAllMocks`.

# Tests

Current tests worked.