feat(backend): save testent toggle in user settings (#5281)
# Motivation

We want to start saving some networks settings in the backend. For
example if we enable or disabled testnets, or what are the enabled
networks.

As first step we create the endpoint to toggle testnets (show them or
hide them).

# Changes

- Expand `Settings` interface to have a network prop (which itself
contains testnets props).
- Create `user_profile` method to save the testnet toggle: true to show
testnets and false to hide them.
- Create endpoint `save_user_testnets_toggle` to open the method
created.\
- Adapt frontend for the newly generated bindings.

# Tests

Included some tests for the new endpoint (in another PR I will split the
settings tests, for better manageability).

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Max Murphy <max@winning.black>