fix(devops): Always run final jobs (#5323)
# Motivation
Many of our workflows have a final job that checks whether all the other
jobs succeeded. This makes it much easier to manage which workflows need
to pass for a PR to merge, than if all the individual jobs need to be
listed in the ruleset.

However I noticed that many of the workflows omit the `always()`
conditional on this final job. Omitting always will make that final step
pointless: If a job dependency, listed in `needs`, is skipped or fails,
the dependent job will be skipped, and a skipped job is considered as
not applicable when considering a merge rule.

Having PRs mergeable when required jobs fail is of course a bug.

# Changes
- Add `always()` as a conditional for all the final steps.

# Tests

---------

Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>