feat(frontend): enables to customize onClose of ManageTokens/ManageTokensModal (#5168)
# Motivation

Bug:
A user that has a token in the token list, but the token is disabled.
Clicking on a direct link to this token (e.g
https://fe3.oisy.com/transactions/?token=ckEURC&network=ICP) will lead
to an infinite loading state.

Fix:
Instead of falling into an infinite loading state, the manage token
modal should be displayed to give the user the option to manually enable
the token.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>