feat(frontend): use bigint instead of BigNumber in util formatToken (#5320)
# Motivation

We need to migrate to `ethers` v6: one of the biggest change was the
deprecation of `BigNumber` in favour of built-in `bigint` (see
[documentation](https://docs.ethers.org/v6/migrating/#migrate-bigint)).

To smooth the migration, we start replacing `BigNumber` with `bigint`
where possible, starting from a very used util that is `formatToken`.

# Changes

- Make `formatToken` accept `bigint` instead `BigNumber`.
- Adapt the usage all throughout the code (quite a lot of changes).

# Tests

Existing tests were adapted and they works.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Max <bitdivine@users.noreply.github.com>