feat(frontend): balance stores will use bigint instead of BigNumber (#5342)
# Motivation

We need to migrate to `ethers` v6: one of the biggest change was the
deprecation of `BigNumber` in favour of built-in `bigint` (see
[documentation](https://docs.ethers.org/v6/migrating/#migrate-bigint)).

After PR https://github.com/dfinity/oisy-wallet/pull/5320, to smooth the
migration, we replace `BigNumber` with `bigint` in the balance store
(and in its usages).

# Changes

- Make type `Balance` as `bigint` instead of `BigNumber`.
- Use type `Balance` in the data for `balanceStore`.
- Adapt the rest of the usages.

# Tests

Existing tests were adapted and they works.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Max <bitdivine@users.noreply.github.com>