feat(backend): add map of Networks settings to user's profile interface (#5313)
# Motivation

We would like to save some networks settings in the backend, for example
if a network is active or not.

As first step, we just expand the structure of settings to include a map
of networks with their respective settings.

After that, we will create a function to set it.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>