feat(frontend): validation in component TokenInput will use bigint instead of BigNumber (#5360)
# Motivation

We need to migrate to `ethers` v6: one of the biggest change was the
deprecation of `BigNumber` in favour of built-in `bigint` (see
[documentation](https://docs.ethers.org/v6/migrating/#migrate-bigint)).

To smooth the migration, we replace `BigNumber` with `bigint` in the
validation props of component `TokenInput`.

# Changes

- Make function props `customValidate` and `customErrorValidate` accept
`bigint` instead of `BigNumber`.
- Adapt the rest of the usages.

# Tests

Existing tests were adapted and they works.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>