feat(frontend): explicit type of method getTransaction of Solana RPC (#5371)
# Motivation

We want to have a more explicit type for the getTransaction method of
Solana RPC.

However, according to [Solana dev
forum](https://solana.stackexchange.com/questions/17643/how-to-infer-or-find-basic-types-for-return-values-in-solana-web3-js-v2-library/17673#17673),
exporting ‘result types’ from `@solana/kit` is impractical, because the
RPC is fluid to the point where its result types change depending on the
particular inputs to the calls themselves.

So, we must first create a sub-function and extract the type from there,
in a more fluid way.

# Changes

- Create function that just wraps the `getTransaction` method of the
Solana RPC.
- Extract the type for such type.
- Use it in the code.

# Tests

Adapted a few mock, the rest is sufficient.