feat(frontend): component ExchangeAmountDisplay will use bigint instead of BigNumber (#5346)
# Motivation

We need to migrate to `ethers` v6: one of the biggest change was the
deprecation of `BigNumber` in favour of built-in `bigint` (see
[documentation](https://docs.ethers.org/v6/migrating/#migrate-bigint)).

After PR https://github.com/dfinity/oisy-wallet/pull/5342, to smooth the
migration, we replace `BigNumber` with `bigint` in the component
`ExchangeAmountDisplay`.

# Changes

- Make component `ExchangeAmountDisplay` accept `amount` as `bigint`
instead of `BigNumber`.
- Adapt the rest of the usages.

# Tests

Existing tests were adapted and they works. Added tests for
`ExchangeAmountDisplay`.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>