feat(frontend): Fee context will use bigint instead of BigNumber (#5381)
# Motivation

We need to migrate to `ethers` v6: one of the biggest change was the
deprecation of `BigNumber` in favour of built-in `bigint` (see
[documentation](https://docs.ethers.org/v6/migrating/#migrate-bigint)).

To smooth the migration, we replace `BigNumber` with `bigint` in the fee
context.

# Changes

- Make interface `TransactionFeeData ` accept `amount` as `bigint`
instead of `BigNumber`. NOTE: when we will migrate to `ethers` v6 we can
switch back to using the built-in `FeeData` interface.
- Transform the results of provider method `getFeeData` into `bigint`.
NOTE: it will not be necessary anymore when we migrate to `ethers` v6.
- Adapt the rest of the usages.

# Tests

Existing tests were sufficient.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>