feat(frontend): implement ModalTokensList component (#5416)
# Motivation

This PR includes a new component that is going to be used in Send and
Swap modals (potentially, also in ManageTokens, bigger refactoring will
be needed there) for displaying clickable token cards. It is based on
recently implement `LogoButton` component. Also, one of the next steps
will be to extend ModalTokensList with a network selector.

<img width="555" alt="Screenshot 2025-03-24 at 20 10 32"
src="https://github.com/user-attachments/assets/a5165dcf-e3a3-4b88-9459-3c7248184787"
/>