feat(backend): methods to save the networks list settings (#5355)
# Motivation

We will start to save the enabled/disabled networks in the user's
profile settings.

So, we provide two methods in the backend to do that: one to update
existing networks settings and the other to overwrite them.

# Changes

- Create function `with_networks` that overwrite or updates (depending
on the input flag) the user's profile networks settings.
- Create method `set_network_settings` that only overwrite the networks
settings.
- Create method `update_network_settings` that only merges the input
values with the existing networks settings.
- Provide both methods as public endpoints.

# Tests

Provided new tests for the new endpoints.

UNRELATED: I split the tests for `testnets` settings from the ones from
`networks` settings for better manageability.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: daviddecentage <80264282+daviddecentage@users.noreply.github.com>