fix(frontend): adjust fallback timeout (#5400)
# Motivation

In some cases the `fallback timeout` of 5 seconds might not be enough.
If a user with a smartphone and a poor internet connection is navigating
to the `token view`, it may take longer than 5 seconds to load the
token. In this cases it is possible that the `manage token modal` opens
before the token is successfully loaded. To reduce the changes of this
happening, we have updated the timeout to 10 seconds.