feat(frontend): Allow dropdown popover to be displayed as modal on small screens (#5434)
# Motivation

Due to the upcoming change in the network dropdown we change th
component slightly. We can pass `asModalOnMobile` prop so that the
dropdown popover renders as a modal on small screens.

# Changes

Changed the way the popover renders and adjusted the network selector
accordingly.

# Tests

On desktop with asModalOnMobile

![image](https://github.com/user-attachments/assets/6e61caaf-9ded-4ced-9fab-63af84939536)

Without the prop

![image](https://github.com/user-attachments/assets/95e8653d-2e13-422b-a7ca-b47724c424b1)

On smaller screen with asModalOnMobile

![image](https://github.com/user-attachments/assets/02cf3ffa-c9a1-4d2d-8427-aab7f2d62646)

Without the prop

![image](https://github.com/user-attachments/assets/70d1aa1e-6416-49d3-97e8-b211795dc63b)


On very small screen with asModalOnMobile

![image](https://github.com/user-attachments/assets/9fdfb6c6-e15b-4ff0-9280-c028aeec3123)

Without the prop

![image](https://github.com/user-attachments/assets/4df11c93-032e-436e-aa9b-360e9ba858d2)

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>