refactor(frontend): no type-derivation for TransactionUiCommon (#5408)
# Motivation

The interface TransactionUiCommon should not depend on the transaction
type of ethers package. It should be independent inside OISY since it
involves all possible networks. So we explicitly define the `from`, `to`
and `blockNumber` properties.

Note that `to` may be undefined for transactions that are approvals, or
that do not require a counterparty.