fix(frontend): scrollable content of ModalTokensList (#5461)
# Motivation

I noticed today that on some screen sizes the TokensList modal displays
with unwanted scrollbar. It appeared to be an issue with the `stretch`
class that was giving this unexpected effect to the list. The solution
is to not use `ContentWithToolbar` in this particular case. Also, I
cleaned up the component by removing some unnecessary HTML/CSS.

<img width="586" alt="Screenshot 2025-03-26 at 16 46 22"
src="https://github.com/user-attachments/assets/436ea7b2-4c78-49e5-9c85-329b019182f4"
/>