feat(e2e): completely hide the carousel (#5401)
# Motivation

The carousel is difficult to handle for the E2E tests, causing quite a
bit of flakiness for each change that we do. That makes the E2E tests
unreliable, and they lose their importance.

Since we need them more than ever, for now we just `mask` the carousel,
looking for stable tests.

We will work to have a permanent state of the carousel as priority in
another PR.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>