feat(frontend): store to check if ICP network is enabled (#5468)
# Motivation

We will start to enable and disable network according to the user
preference. That means that we could have the Internet Computer network
disabled.

To that scope we need to turn off some features that would not be
possible if the Internet Computer network is turned off.

# Changes

- Create derived store to check if Internet Computer network is enabled.
- Use the store to disable buttons that convert CK tokens.
- Use the store to hide the ICP addresses.

# Next Steps

- Use the store to hide the ICP/ICRC tokens

# Tests

I simulated the disabled network:



https://github.com/user-attachments/assets/067d2cb1-6a14-4cb8-bb36-9f9ad262527f