feat(frontend): create feature flag to avoid using user networks (#5486)
# Motivation

We will start to enable/disable networks based on the preferences in the
user profile. However, we want to have more control on this new feature.
So we create a flag that would default the user networks the value of
all mainnets networks enabled, and testnets too if they are enabled with
the toggle. This is as it is before the feautre.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>