feat(frontend): make NetworksSwitcher re-usable for tokens list filters (#5484)
# Motivation

The idea of this refactoring/update is to make NetworksSwitcher
re-usable for tokens list filters. To do so, we need to make the
component agnostic of what happens next after a network has been
selected. This logic is now moved to consumers (`Header` in case of this
PR, `ModalTokensList` - my next step).

---------

Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>