feat(frontend): stores to check if Ethereum networks are enabled (#5476)
# Motivation

We will start to enable and disable network according to the user
preference. That means that we could have the Ethereum networks
disabled.

To that scope we need to turn off some features that would not be
possible if the networks are turned off.

# Changes

- Create derived stores to check if Ethereum networks are enabled: one
for Mainnet and one for Sepolia.
- Use the stores to hide buttons that convert CK tokens.
- Use the stores to hide the Ethereum address.
- Use the stores to hide the Info modal.

# Next Steps

- Use the store to hide the Ethereum tokens (native and ERC20).

# Tests

I simulated the disabled network:




https://github.com/user-attachments/assets/1b175479-8ebf-4ac9-b046-3ed6fd60430c